updates for feedback
