+++ components/nimbus/tests/test_restart.rs
@@ -2,7 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+#[cfg(feature = "rkv-safe-mode")]
 mod common;
+
 #[cfg(feature = "rkv-safe-mode")]
 #[cfg(test)]
 mod test {
